feat: expose snip-12 encoding api
